requirements.txt
azureml-sdk
azure-storage-blob
joblib
pandas
requests
scikit-learn

-e src